\name{seq.hotSPOT}
\alias{seq.hotSPOT}

\title{Targeted sequencing panel design based on mutation hotspots}

\description{
seq.hotSPOT provides a resource for designing effective sequencing panels to help
improve mutation capture efficacy for ultradeep sequencing projects. Using SNV datasets,
this package designs custom panels for any tissue of interest and identify the genomic regions
likely to contain the most mutations. Establishing efficienttargeted sequencing panels can
allow researchers to study mutation burden in tissues at high depth without the
economic burden of whole-exome or whole-genome sequencing. This tool was developed to make
high-depth sequencing panels to study low-frequency clonal mutations in clinically normal
and cancerous tissues.
}
\value{
A package containing functions which generate optimal targeted sequencing panels based on mutation hotspots.
}
